database/requirements.txt
openai
numpy
faiss-cpu
